refactor(yacfg): yacfg_cli refactored
- Import the `metadata` module from `importlib_metadata` if the Python version is less than 3.10, and from `importlib` otherwise.
- Adjust file structure by renaming `src/yacfg/cli_arguments.py` to `src/yacfg/cli/cli_arguments.py` and `src/yacfg/yacfg_cli.py` to `src/yacfg/cli/yacfg_cli.py`.
- Improve and fix issues in the `yacfg_cli` module related to logging settings, command-line option handling, and calling appropriate functions based on the provided arguments.
