+++ tests/test_yacfg/cli_arguments/test_boolize.py
@@ -14,7 +14,7 @@
 
 import pytest
 
-import yacfg.cli_arguments
+import yacfg.cli.cli_arguments
 
 dataset_boolize = (
     (None, None),
@@ -37,9 +37,9 @@ dataset_boolize = (
 
 @pytest.mark.parametrize("value,result", dataset_boolize)
 def test_boolize_values(value, result):
-    assert yacfg.cli_arguments.boolize(value) == result
+    assert yacfg.cli.cli_arguments.boolize(value) == result
 
 
 def test_boolize_exception():
     with pytest.raises(ValueError):
-        yacfg.cli_arguments.boolize("test")
+        yacfg.cli.cli_arguments.boolize("test")
